Structure & Refactor ITS Frontend
