+++ frontend/src/components/its/ClustersTable/dialogs/DetachmentLogsDialog.tsx
