+++ frontend/src/pages/Dashboard.tsx
@@ -32,7 +32,7 @@ import { useBPQueries } from '../hooks/queries/useBPQueries';
 import { useTranslation } from 'react-i18next';
 
 // Lazy load the ClusterDetailDialog component
-const ClusterDetailDialog = lazy(() => import('../components/ClusterDetailDialog'));
+const ClusterDetailDialog = lazy(() => import('../components/its/ClustersTable/dialogs/ClusterDetailDialog'));
 
 // Health indicator component
 const HealthIndicator = ({ value }: { value: number }) => {
